+++ Week9/9-1/src/components/Cart/CartList/CartList.tsx
@@ -4,15 +4,13 @@ import { useSelector } from "../../../hooks/useCustomRedux";
 const CartList = () => {
   const { cartItems } = useSelector((state) => state.cart);
   return (
-    <>
-      <div className="flex flex-col items-center justify-center">
-        <ul>
-          {cartItems.map((item) => (
-            <CartItem key={item.id} lp={item} />
-          ))}
-        </ul>
-      </div>
-    </>
+    <div className="flex flex-col items-center justify-center">
+      <ul>
+        {cartItems.map((item) => (
+          <CartItem key={item.id} lp={item} />
+        ))}
+      </ul>
+    </div>
   );
 };
