+++ Week9/9-1/src/slices/Cart/cartSlice.types.ts
@@ -1,0 +1,7 @@
+import type { CartItems } from "../../types/cart";
+
+export interface CartState {
+  cartItems: CartItems;
+  amount: number;
+  total: number;
+}
