+++ Week9/9-1/src/slices/Modal/modalSlice.ts
@@ -1,8 +1,5 @@
 import { createSlice } from "@reduxjs/toolkit";
-
-interface ModalState {
-  isOpen: boolean;
-}
+import type { ModalState } from "./modalSlice.types";
 
 const initialState: ModalState = {
   isOpen: false,
