+++ Week9/9-1/src/slices/Modal/modalSlice.types.ts
@@ -1,0 +1,3 @@
+export interface ModalState {
+  isOpen: boolean;
+}
